requirements.txt
requests==2.27.1
beautifulsoup4
streamlit
numpy
Pillow
flask
keras~=2.6.0
tensorflow~=2.6.0